+++ examples/children/vite/src/main.js
@@ -18,7 +18,6 @@ const routes = [
 // app.use(router)
 // app.mount('#vite-app')
 
-
 let app = null
 let router = null
 let history = null
@@ -35,12 +34,28 @@ function mount () {
   app.mount('#vite-app')
 
   console.log('微应用child-vite渲染了')
+
+  // eventCenterForVite 是基座添加到window的数据通信对象
+  // 主动获取基座下发的数据
+  console.log('child-vite getData:', window.eventCenterForVite?.getData())
+
+  // 监听基座下发的数据变化
+  window.eventCenterForVite?.addDataListener((data) => {
+    console.log('child-vite addDataListener:', data)
+  })
+
+  // 向基座发送数据
+  setTimeout(() => {
+    window.eventCenterForVite?.dispatch({ myname: 'child-vite' })
+  }, 3000)
 }
 
 // 将卸载操作放入 unmount 函数
 function unmount () {
   app?.unmount()
   history?.destroy()
+  // 卸载所有数据监听函数
+  window.eventCenterForVite?.clearDataListener()
   app = null
   router = null
   history = null