+++ src/index.ts
@@ -1,3 +1,4 @@
-export { default as preFetch } from './prefetch'
 export { default } from './micro_app'
+export { default as preFetch } from './prefetch'
 export { removeDomScope, pureCreateElement, version, setCurrentAppName } from './libs/utils'
+export { EventCenterForMicroApp } from './interact'